% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-handler.R
\name{extractYaml}
\alias{extractYaml}
\title{Get YAML header from a text vector}
\usage{
extractYaml(text)
}
\arguments{
\item{text}{a vector of string the text read from a file}
}
\value{
a list representing the yaml that can be exported back to yaml format
}
\description{
Get YAML header from a text vector
}
