% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-handler.R
\name{getYamlStartEnd}
\alias{getYamlStartEnd}
\title{Identify where the yaml header starts and end in a file}
\usage{
getYamlStartEnd(text, only_content = TRUE)
}
\arguments{
\item{text}{a string vector containing the read lines of a file}

\item{only_content}{a logical defining if the returned index should be for
the yaml delimiters or the yaml content.}
}
\value{
a numeric vector of size two representing the index
}
\description{
Identify where the yaml header starts and end in a file
}
