% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\title{Create a new report}
\usage{
new_report(file_name, path = NA)
}
\arguments{
\item{file_name}{a string that represent the name of the report that will be created.}

\item{path}{a string that represent the path where to create the report.}
}
\description{
Create a new report
}
\examples{
new_report("my_report", "report_folder/")
}
