% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-report.R
\name{newReport}
\alias{newReport}
\title{Create a new report}
\usage{
newReport(
  report_title,
  path = "Reports/",
  template = "report",
  subtitle = NA,
  author = NA,
  datetime = NA
)
}
\arguments{
\item{path}{a string that represent the path where to create the report.}

\item{template}{one string (either \code{report} or \code{tutorial}) representing the type of template to initialize.}

\item{subtitle}{the subtitle of the report. empty by default.}

\item{author}{the name of the author(s). "esqlabs Gmbh" by default.}

\item{datetime}{the date time that will appear on the report in "YYYY-MM-DD HH:MM:SS" format. default to time of generation of the report.}

\item{file_name}{a string that represent the name of the report that will be created.}
}
\description{
Create a new report
}
\examples{
newReport("my_report", "report_folder/")
}
