% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-handler.R
\name{replaceYaml}
\alias{replaceYaml}
\title{Replace yaml header of a file}
\usage{
replaceYaml(file_path, new_yaml)
}
\arguments{
\item{file_path}{a string representing the file containing the yaml header to replace.}

\item{new_yaml}{a list representing the yaml header to inject}
}
\description{
Replace yaml header of a file
}
